requirements-test.txt
pytest==7.2.1
pytest-asyncio==0.23.7
pytest-order==1.1.0
pytest-env==1.1.3
beautifulsoup4
fastapi==0.111.0
# faiss_cpu==1.8.0
faiss_cpu==1.8.0
pillow==10.3.0
psycopg2-binary==2.9.9
torchvision==0.15
open-metric-learning==2.0.0
pandas
packaging